♻️ refactor(document): relocate Tesseract.js integration
